+++ src/components/layouts/Layout2.jsx
@@ -1,10 +1,10 @@
 import Header from '../headers/Header2';
 
-//Layout Padrão com Header simples
-function Layout() {
+//Layout Padrão com Header completo
+function Layout({ placeholder }) {
   return (
     <>
-      <Header />
+      <Header placeholder={placeholder} />
     </>
   );
 }